noname::backends

Trait Backend

pub trait Backend: Clone {
    type Field: BackendField;
    type Var: BackendVar;
    type GeneratedWitness;

    // Required methods
    fn poseidon() -> FnHandle<Self>;
    fn new_internal_var(&mut self, val: Value<Self>, span: Span) -> Self::Var;
    fn neg(&mut self, var: &Self::Var, span: Span) -> Self::Var;
    fn add(&mut self, lhs: &Self::Var, rhs: &Self::Var, span: Span) -> Self::Var;
    fn add_const(
        &mut self,
        var: &Self::Var,
        cst: &Self::Field,
        span: Span,
    ) -> Self::Var;
    fn mul(&mut self, lhs: &Self::Var, rhs: &Self::Var, span: Span) -> Self::Var;
    fn mul_const(
        &mut self,
        var: &Self::Var,
        cst: &Self::Field,
        span: Span,
    ) -> Self::Var;
    fn assert_eq_const(&mut self, var: &Self::Var, cst: Self::Field, span: Span);
    fn assert_eq_var(&mut self, lhs: &Self::Var, rhs: &Self::Var, span: Span);
    fn add_public_input(&mut self, val: Value<Self>, span: Span) -> Self::Var;
    fn add_private_input(&mut self, val: Value<Self>, span: Span) -> Self::Var;
    fn add_public_output(&mut self, val: Value<Self>, span: Span) -> Self::Var;
    fn add_constant(
        &mut self,
        label: Option<&'static str>,
        value: Self::Field,
        span: Span,
    ) -> Self::Var;
    fn compute_var(
        &self,
        env: &mut WitnessEnv<Self::Field>,
        var: &Self::Var,
    ) -> Result<Self::Field>;
    fn finalize_circuit(
        &mut self,
        public_output: Option<Var<Self::Field, Self::Var>>,
        returned_cells: Option<Vec<Self::Var>>,
        disable_safety_check: bool,
    ) -> Result<()>;
    fn generate_witness(
        &self,
        witness_env: &mut WitnessEnv<Self::Field>,
        sources: &Sources,
        typed: &Mast<Self>,
    ) -> Result<Self::GeneratedWitness>;
    fn generate_asm(&self, sources: &Sources, debug: bool) -> String;
    fn log_var(&mut self, var: &VarInfo<Self::Field, Self::Var>, span: Span);

    // Provided methods
    fn init_circuit(&mut self) { ... }
    fn sub(&mut self, lhs: &Self::Var, rhs: &Self::Var, span: Span) -> Self::Var { ... }
    fn compute_val(
        &self,
        env: &mut WitnessEnv<Self::Field>,
        val: &Value<Self>,
        cache_key: usize,
    ) -> Result<Self::Field> { ... }
    fn print_log(
        &self,
        witness_env: &mut WitnessEnv<Self::Field>,
        logs: &[(Span, VarInfo<Self::Field, Self::Var>)],
        sources: &Sources,
        typed: &Mast<Self>,
    ) -> Result<()> { ... }

}

Required Associated Types

type Field: BackendField

The circuit field / scalar field that the circuit is written on.

type Var: BackendVar

The CellVar type for the backend. Different backend is allowed to have different CellVar types.

type GeneratedWitness

The generated witness type for the backend. Each backend may define its own witness format to be generated.

Required Methods

fn poseidon() -> FnHandle<Self>

poseidon crypto builtin function for different backends

fn new_internal_var(&mut self, val: Value<Self>, span: Span) -> Self::Var

Create a new cell variable and record it. It increments the variable index for look up later.

fn neg(&mut self, var: &Self::Var, span: Span) -> Self::Var

negate a var

fn add(&mut self, lhs: &Self::Var, rhs: &Self::Var, span: Span) -> Self::Var

add two vars

fn add_const( &mut self, var: &Self::Var, cst: &Self::Field, span: Span, ) -> Self::Var

add a var with a constant

fn mul(&mut self, lhs: &Self::Var, rhs: &Self::Var, span: Span) -> Self::Var

multiply a var with another var

fn mul_const( &mut self, var: &Self::Var, cst: &Self::Field, span: Span, ) -> Self::Var

multiply a var with a constant

fn assert_eq_const(&mut self, var: &Self::Var, cst: Self::Field, span: Span)

add a constraint to assert a var equals a constant

fn assert_eq_var(&mut self, lhs: &Self::Var, rhs: &Self::Var, span: Span)

add a constraint to assert a var equals another var

fn add_public_input(&mut self, val: Value<Self>, span: Span) -> Self::Var

Process a public input

fn add_private_input(&mut self, val: Value<Self>, span: Span) -> Self::Var

Process a private input

fn add_public_output(&mut self, val: Value<Self>, span: Span) -> Self::Var

Process a public output

fn add_constant( &mut self, label: Option<&'static str>, value: Self::Field, span: Span, ) -> Self::Var

This should be called only when you want to constrain a constant for real. Gates that handle constants should always make sure to call this function when they want them constrained.

fn compute_var( &self, env: &mut WitnessEnv<Self::Field>, var: &Self::Var, ) -> Result<Self::Field>

Backends should implement this function to load and compute the value of a CellVar.

fn finalize_circuit( &mut self, public_output: Option<Var<Self::Field, Self::Var>>, returned_cells: Option<Vec<Self::Var>>, disable_safety_check: bool, ) -> Result<()>

Finalize the circuit by doing some sanitizing checks.

fn generate_witness( &self, witness_env: &mut WitnessEnv<Self::Field>, sources: &Sources, typed: &Mast<Self>, ) -> Result<Self::GeneratedWitness>

Generate the witness for a backend.

fn generate_asm(&self, sources: &Sources, debug: bool) -> String

Generate the asm for a backend.

fn log_var(&mut self, var: &VarInfo<Self::Field, Self::Var>, span: Span)

Provided Methods

fn init_circuit(&mut self)

Init circuit

fn sub(&mut self, lhs: &Self::Var, rhs: &Self::Var, span: Span) -> Self::Var

sub two vars

fn compute_val( &self, env: &mut WitnessEnv<Self::Field>, val: &Value<Self>, cache_key: usize, ) -> Result<Self::Field>

fn print_log( &self, witness_env: &mut WitnessEnv<Self::Field>, logs: &[(Span, VarInfo<Self::Field, Self::Var>)], sources: &Sources, typed: &Mast<Self>, ) -> Result<()>

print the log given the log_info

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.