Trait noname::backends::BackendVar

pub trait BackendVar: Clone + Debug + PartialEq + Eq { }

This trait allows different backends to have different cell var types. It is intended to make it opaque to the frondend.

Object Safety

This trait is not object safe.