noname::circuit_writer
Module fn_env

Structs

FnEnv

Is used to store functions’ scoped variables. This include inputs and output of the function, as well as local variables. You can think of it as a function call stack.

VarInfo

Information about a variable.