Struct noname::circuit_writer::CircuitWriter

pub struct CircuitWriter { /* private fields */ }

Implementations

impl CircuitWriter

pub fn generate_asm(&self, sources: &Sources, debug: bool) -> String

impl CircuitWriter

pub fn compiled_gates(&self) -> &[Gate]

Returns the compiled gates of the circuit.

pub fn new_internal_var(&mut self, val: Value, span: Span) -> CellVar

pub fn compute_constant(&self, var: CellVar, span: Span) -> Result<Field>

pub fn num_gates(&self) -> usize

pub fn add_constant(
&mut self,
label: Option<&'static str>,
value: Field,
span: Span
) -> CellVar

This should be called only when you want to constrain a constant for real. Gates that handle constants should always make sure to call this function when they want them constrained.

pub fn add_gate(
&mut self,
note: &'static str,
typ: GateKind,
vars: Vec<Option<CellVar>>,
coeffs: Vec<Field>,
span: Span
)

creates a new gate, and the associated row in the witness/execution trace.

pub fn add_public_inputs(&mut self, name: String, num: usize, span: Span) -> Var

pub fn add_public_outputs(&mut self, num: usize, span: Span)

pub fn add_private_inputs(&mut self, name: String, num: usize, span: Span) -> Var

impl CircuitWriter

pub fn expr_type(&self, expr: &Expr) -> Option<&TyKind>

pub fn node_type(&self, node_id: usize) -> Option<&TyKind>

pub fn struct_info(&self, qualified: &FullyQualified) -> Option<&StructInfo>

pub fn fn_info(&self, qualified: &FullyQualified) -> Option<&FnInfo>

pub fn const_info(&self, qualified: &FullyQualified) -> Option<&ConstInfo>

pub fn size_of(&self, typ: &TyKind) -> usize

pub fn add_local_var(
&self,
fn_env: &mut FnEnv,
var_name: String,
var_info: VarInfo
)

pub fn get_local_var(&self, fn_env: &FnEnv, var_name: &str) -> VarInfo

pub fn main_info(&self) -> Result<&FnInfo>

Retrieves the FnInfo for the main() function. This function should only be called if we know there’s a main function, if there’s no main function it’ll panic.

pub fn error(&self, kind: ErrorKind, span: Span) -> Error

impl CircuitWriter

pub fn generate_circuit(
typed: TypeChecker,
double_generic_gate_optimization: bool
) -> Result<CompiledCircuit>