noname::cli::manifest

Function read_manifest

pub fn read_manifest(path: &Utf8PathBuf) -> Result<Manifest>

This retrieves a dependency listed in the manifest file. It downloads it from github, and stores it under the deps directory. A dependency is expected go be given as “user/repo”.