Function noname::cli::packages::get_dep

pub fn get_dep(dep: &Library) -> Result<Manifest>

This retrieves a dependency listed in the manifest file. It downloads it from github, and stores it under the deps directory. A dependency is expected go be given as “user/repo”. Note that this does not download the dependencies of the dependency.