Module noname::compiler

This module is a wrapper API around noname. It is important that user-facing features use the functions here, as they attach the correct filename and source code to errors. It does that by transforming our [Error] type into a [miette::Error] type for all functions here. (via the IntoMiette trait that we define here.)

Structs

Sources

Contains the association between a counter and the corresponding filename and source code.

Traits

IntoMiette

Functions

compile

generate_witness

get_nast

typecheck_next_file

typecheck_next_file_inner

This should not be used directly. Check [get_tast] instead.