Enum noname::error::ErrorKind

pub enum ErrorKind {

    UnexpectedError(&'static str),
    AssignmentToImmutableVariable,
    UnknownDependency(String),
    UnknownExternalFn(String, String),
    UnknownExternalStruct(String, String),
    MainFunctionInLib,
    ShadowingBuiltIn(String),
    ParsingError(ParsingError),
    ConstArgumentNotForMain,
    InvalidFieldAccessExpression,
    NotAStaticMethod,
    MismatchFunctionArguments(usize, usize),
    OrderOfConstDeclaration,
    OrderOfUseDeclaration,
    MissingParenthesis,
    PubArgumentOutsideMain,
    RecursiveMain,
    InvalidToken,
    MissingType,
    MissingToken,
    ExpectedToken(TokenKind),
    InvalidPath(&'static str),
    InvalidEndOfLine,
    InvalidFunctionSignature(&'static str),
    InvalidFunctionName,
    InvalidTypeName,
    InvalidType,
    ReservedType(String),
    InvalidArraySize,
    InvalidField(String),
    InvalidRangeSize,
    InvalidStatement,
    MissingExpression,
    InvalidExpression,
    InvalidIdentifier(String),
    InvalidFnCall(&'static str),
    UseAfterFn,
    WrongArgumentType {
        fn_name: String,
        arg_name: String,
        expected_ty: String,
        observed_ty: String,
    },
    CannotComputeExpression,
    MismatchType(TyKind, TyKind),
    UndefinedVariable,
    ArgumentTypeMismatch(TyKind, TyKind),
    FunctionReturnsType(String),
    MissingPublicArg(String),
    MissingPrivateArg(String),
    CannotConvertToField(String),
    MissingReturn,
    NoReturnExpected,
    SelfHasAttribute,
    ReturnTypeMismatch(TyKind, TyKind),
    UnexpectedReturn,
    StdImport(String),
    DuplicateModule(String),
    PublicOutputReserved(String),
    UndefinedFunction(String),
    UndefinedModule(String),
    InvalidAttribute(AttributeKind),
    UnusedReturnValue,
    ArrayIndexOutOfBounds(usize, usize),
    NoOneLetterVariable,
    ExpectedConstant,
    KimchiSetup(SetupError),
    KimchiProver(ProverError),
    KimchiVerifier(VerifyError),
    InvalidWitness(usize),
    UnusedInput(String),
    PrivateInputNotUsed,
    DuplicateDefinition(String),
    InvalidAssignmentExpression,
    NoArgsInMain,
    LocalVariableNotFound(String),
    ConstantInOutput,
    MismatchStructFields(String),
    InvalidStructField(String, String),
    InvalidStructFieldType(TyKind, TyKind),
    MethodCallOnNonCustomStruct,
    ArrayAccessOnNonArray,
    UndefinedStruct(String),
    UndefinedField(String, String),
    AssertionFailed,
    InvalidConstType,
    NoMainFunction,

}

The type of error.

Variants

UnexpectedError(&'static str)

this error is for testing. You can use it when you want to quickly see in what file and what line of code the error is.

AssignmentToImmutableVariable

UnknownDependency(String)

UnknownExternalFn(String, String)

UnknownExternalStruct(String, String)

MainFunctionInLib

ShadowingBuiltIn(String)

ParsingError(ParsingError)

ConstArgumentNotForMain

InvalidFieldAccessExpression

NotAStaticMethod

MismatchFunctionArguments(usize, usize)

OrderOfConstDeclaration

OrderOfUseDeclaration

MissingParenthesis

PubArgumentOutsideMain

RecursiveMain

InvalidToken

MissingType

MissingToken

ExpectedToken(TokenKind)

InvalidPath(&'static str)

InvalidEndOfLine

InvalidFunctionSignature(&'static str)

InvalidFunctionName

InvalidTypeName

InvalidType

ReservedType(String)

InvalidArraySize

InvalidField(String)

InvalidRangeSize

InvalidStatement

MissingExpression

InvalidExpression

InvalidIdentifier(String)

InvalidFnCall(&'static str)

UseAfterFn

WrongArgumentType

Fields

fn_name: String

arg_name: String

expected_ty: String

observed_ty: String

CannotComputeExpression

MismatchType(TyKind, TyKind)

UndefinedVariable

ArgumentTypeMismatch(TyKind, TyKind)

FunctionReturnsType(String)

MissingPublicArg(String)

MissingPrivateArg(String)

CannotConvertToField(String)

MissingReturn

NoReturnExpected

SelfHasAttribute

ReturnTypeMismatch(TyKind, TyKind)

UnexpectedReturn

StdImport(String)

DuplicateModule(String)

PublicOutputReserved(String)

UndefinedFunction(String)

UndefinedModule(String)

InvalidAttribute(AttributeKind)

UnusedReturnValue

ArrayIndexOutOfBounds(usize, usize)

NoOneLetterVariable

ExpectedConstant

KimchiSetup(SetupError)

KimchiProver(ProverError)

KimchiVerifier(VerifyError)

InvalidWitness(usize)

UnusedInput(String)

PrivateInputNotUsed

DuplicateDefinition(String)

InvalidAssignmentExpression

NoArgsInMain

LocalVariableNotFound(String)

ConstantInOutput

MismatchStructFields(String)

InvalidStructField(String, String)

InvalidStructFieldType(TyKind, TyKind)

MethodCallOnNonCustomStruct

ArrayAccessOnNonArray

UndefinedStruct(String)

UndefinedField(String, String)

AssertionFailed

InvalidConstType

NoMainFunction