Struct noname::error::Error

pub struct Error {
    pub label: &'static str,
    pub kind: ErrorKind,
    pub span: Span,
}

An error in noname.

Fields
label: &'static str

A hint as to where the error happened (e.g. type-checker, parser, lexer, etc.)

kind: ErrorKind

The type of error.

span: Span

Indicate where the error occurred in the source code.

Implementations

impl Error

pub fn new(label: &'static str, kind: ErrorKind, span: Span) -> Self

Creates a new [Error] from an ErrorKind.