Function noname::mast::instantiate_fn_call

pub fn instantiate_fn_call<B: Backend>(
    ctx: &mut MastCtx<B>,
    fn_info: FnInfo<B>,
    args: &[ExprMonoInfo],
    span: Span,
) -> Result<(FnInfo<B>, Option<TyKind>)>

Overall, the function call checking process is as follows:

resolve generic values from observed function args

propagate the resolved values within the function body

type check the observed return and the resolved expected return