Enum noname::parser::types::TyKind

pub enum TyKind {
    Field,
    Custom {
        module: ModulePath,
        name: String,
    },
    BigInt,
    Array(Box<TyKind>, u32),
    Bool,
}

Variants

Field

The main primitive type. ’Nuf said.

Custom

Custom / user-defined types

Fields

module: ModulePath

name: String

BigInt

This could be the same as Field, but we use this to also track the fact that it’s a constant.

Array(Box<TyKind>, u32)

An array of a fixed size.

Bool

A boolean (true or false).

Implementations

impl TyKind

pub fn match_expected(&self, expected: &TyKind) -> bool

pub fn same_as(&self, other: &TyKind) -> bool