noname::serialization
Module ser

You can use this module for serialization and deserializing arkworks types with serde. Simply use the following attribute on your field: #[serde(with = "o1_utils::serialization::ser") attribute"]

Functions

deserialize

You can use this to deserialize an arkworks type with serde and the “deserialize_with” attribute. See https://serde.rs/field-attrs.html

serialize

You can use this to serialize an arkworks type with serde and the “serialize_with” attribute. See https://serde.rs/field-attrs.html