noname
Module syntax

A number of helper function to check the syntax of some types.

Functions

is_generic_parameter

Returns true if the given string is generic parameter

is_hexadecimal

Returns true if the given string is an hexadecimal string (0x…)

is_identifier

Returns true if the given string is an identifier (starts with a lowercase letter)

is_identifier_or_type

Returns true if the given string is an identifier or type

is_numeric

Returns true if the given string is a number in decimal.

is_type

Returns true if the given string is a type Check camel case