Struct noname::tokens::Tokens

pub struct Tokens {
    pub peeked: Option<Token>,
    /* private fields */
}

Fields
peeked: Option<Token>
Implementations

impl Tokens

pub fn new(tokens: Vec<Token>) -> Self

pub fn peek(&mut self) -> Option<Token>

Peeks into the next token without advancing the iterator.

pub fn bump(&mut self, ctx: &mut ParserCtx) -> Option<Token>

Like next() except that it also stores the last seen token in the given context (useful for debugging)

pub fn bump_err(&mut self, ctx: &mut ParserCtx, err: ErrorKind) -> Result<Token>

Like Self::bump but errors with err pointing to the latest token

pub fn bump_expected(
&mut self,
ctx: &mut ParserCtx,
typ: TokenKind
) -> Result<Token>

Like Self::bump but errors if the token is not typ

pub fn bump_ident(
&mut self,
ctx: &mut ParserCtx,
kind: ErrorKind
) -> Result<Ident>