noname

Module var

Structs

Var

Represents a variable in the noname language, or an anonymous variable during computation of expressions.

Enums

ConstOrCell

Represents a cell in the execution trace.

Value

A variable’s actual value in the witness can be computed in different ways.

VarOrRef

Represents a variable in the circuit, or a reference to one. Note that mutable variables are always passed as references, as one needs to have access to the variable name to be able to reassign it in the environment.

Type Aliases

HintFn

The signature of a hint function