Crate noname

This is a high-level language to write circuits that you can prove in kimchi. Refer to the book for more information.

Modules

backends

circuit_writer

cli

compiler

This module is a wrapper API around noname. It is important that user-facing features use the functions here, as they attach the correct filename and source code to errors. It does that by transforming our [Error] type into a [miette::Error] type for all functions here. (via the IntoMiette trait that we define here.)

constants

constraints

error

helpers

imports

inputs

Used to parse public and private inputs to a program.

lexer

mast

name_resolution

parser

serialization

This adds a few utility functions for serializing and deserializing arkworks types that implement [CanonicalSerialize] and [CanonicalDeserialize].

server

stdlib

syntax

A number of helper function to check the syntax of some types.

type_checker

utils

var

witness