Module noname::backends::r1cs

Modules

builtin

snarkjs

Structs

CellVar

Constraint

an R1CS constraint Each constraint comprises of 3 linear combinations from 3 matrices. It represents a constraint in math: a * b = c.

GeneratedWitness

An intermediate struct for SnarkjsExporter to reorder the witness and convert to snarkjs format.

LinearCombination

Linear combination of variables and constants. For example, the linear combination is represented as a * f_a + b * f_b + f_c f_a and f_b are the coefficients of a and b respectively. a and b are represented by CellVar. The constant f_c is represented by the constant field, will always multiply with the variable at index 0 which is always 1.

R1CS

R1CS backend with bls12_381 field.

Type Aliases

R1csBls12381Field

R1csBn254Field