noname::cli

Module packages

Structs

DependencyGraph

DependencyNode

UserRepo

A dependency is a Github user/repo pair.

Functions

download_from_github

download package from github

get_dep

This retrieves a dependency listed in the manifest file. It downloads it from github, and stores it under the deps directory. A dependency is expected go be given as “user/repo”. Note that this does not download the dependencies of the dependency.

get_dep_code

get_deps_of_package

Returns the dependencies of a package (given it’s manifest).

is_lib

validate_package_and_get_manifest

A valid package must have a valid Noname.toml as well as a lib.no file.