noname::constraints

Module field

Functions

add

Adds two field elements

equal

This takes variables that can be anything, and returns a boolean

if_else

if_else_inner

mul

Multiplies two field elements

neg

Negates a field element

not_equal

Returns 1 if lhs != rhs, 0 otherwise

sub

Subtracts two variables, we only support variables that are of length 1.