noname
Module mast

Modules

ast

Structs

ExprMonoInfo

ExprMonoInfo holds the monomorphized expression node and its resolved type.

MTypeInfo

MTypeInfo holds the resolved type info to pass within a function scope. It is stored in the scope context environment MonomorphizedFnEnv.

Mast

Mast relies on the TAST for the information about the “unresolved” types to monomorphize. Things such as loading the function AST and struct AST from fully qualified names. After monomorphization process, the following data will be updated:

MastCtx

A context to store the last node id for the monomorphized AST.

MonomorphizedFnEnv

A storage to manage the variables in function scopes.

Enums

PropagatedConstant

Functions

error

instantiate_fn_call

Overall, the function call checking process is as follows:

monomorphize

Monomorphize the main function. This is the entry point of the monomorphization process. It stores the monomorphized AST at the end.

monomorphize_block

Monomorphize a block of statements.

monomorphize_stmt

Monomorphize a statement.